# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dashboard ITIS
  description: API for academic performance analytics dashboard
  version: 1.0.0
servers:
- url: http://localhost:8000
paths:
...
  /api/v1/grades:
    get:
      tags:
      - Grades
      summary: Get Grades
      operationId: get_grades_api_v1_grades_get
      security:
      - OAuth2PasswordBearer:
        - grades:list
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_StudentGradeResponse_'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Grades
      summary: Create Grade
      operationId: create_grade_api_v1_grades_post
      security:
      - OAuth2PasswordBearer:
        - grades:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradePublic'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/grades/import:
    post:
      tags:
      - Grades
      summary: Import Grades
      operationId: import_grades_api_v1_grades_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_grades_api_v1_grades_import_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeImportResult'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - OAuth2PasswordBearer:
        - grades:create
  /api/v1/grades/export:
    get:
      tags:
      - Grades
      summary: Export Grades
      operationId: export_grades_api_v1_grades_export_get
      security:
      - OAuth2PasswordBearer:
        - grades:list
      parameters:
      - name: student_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Student Id
      - name: course_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/grades/{grade_id}:
    get:
      tags:
      - Grades
      summary: Get Grade
      operationId: get_grade_api_v1_grades__grade_id__get
      security:
      - OAuth2PasswordBearer:
        - grades:read
      parameters:
      - name: grade_id
        in: path
        required: true
        schema:
          type: integer
          title: Grade Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentGradeResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Grades
      summary: Update Grade
      operationId: update_grade_api_v1_grades__grade_id__put
      security:
      - OAuth2PasswordBearer:
        - grades:update
      parameters:
      - name: grade_id
        in: path
        required: true
        schema:
          type: integer
          title: Grade Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradePublic'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Grades
      summary: Delete Grade
      operationId: delete_grade_api_v1_grades__grade_id__delete
      security:
      - OAuth2PasswordBearer:
        - grades:delete
      parameters:
      - name: grade_id
        in: path
        required: true
        schema:
          type: integer
          title: Grade Id
      responses:
        '204':
          description: Successful Response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/students/{student_id}/grades:
    get:
      tags:
      - Grades
      summary: Get Student Grades
      operationId: get_student_grades_api_v1_students__student_id__grades_get
      security:
      - OAuth2PasswordBearer:
        - grades:list
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_StudentGradeResponse_'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Body_import_grades_api_v1_grades_import_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_import_grades_api_v1_grades_import_post
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - items: {}
            type: array
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorResponse
    GradeCreate:
      properties:
        student_id:
          type: integer
          title: Student Id
        course_id:
          type: integer
          title: Course Id
        score:
          type: number
          maximum: 100.0
          minimum: 0.0
          title: Score
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
      type: object
      required:
      - student_id
      - course_id
      - score
      title: GradeCreate
    GradeImportError:
      properties:
        row:
          type: integer
          title: Row
        message:
          type: string
          title: Message
      type: object
      required:
      - row
      - message
      title: GradeImportError
    GradeImportResult:
      properties:
        created:
          type: integer
          title: Created
        failed:
          type: integer
          title: Failed
        errors:
          items:
            $ref: '#/components/schemas/GradeImportError'
          type: array
          title: Errors
          default: []
      type: object
      required:
      - created
      - failed
      title: GradeImportResult
    GradePublic:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        student_id:
          type: integer
          title: Student Id
        course_id:
          type: integer
          title: Course Id
        score:
          type: number
          maximum: 100.0
          minimum: 0.0
          title: Score
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
      type: object
      required:
      - student_id
      - course_id
      - score
      title: GradePublic
    GradeUpdate:
      properties:
        student_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Student Id
        course_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Course Id
        score:
          anyOf:
          - type: number
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Score
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
      type: object
      title: GradeUpdate
...
    PaginatedResponse_StudentGradeResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/StudentGradeResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - skip
      - limit
      title: PaginatedResponse[StudentGradeResponse]
...
    StudentGradeResponse:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        course_id:
          type: integer
          title: Course Id
        score:
          type: number
          title: Score
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        course_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Course Name
      type: object
      required:
      - id
      - student_id
      - course_id
      - score
      - created_at
      - updated_at
      title: StudentGradeResponse
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          refreshUrl: /api/v1/auth/refresh
          scopes:
            users:list: List users
            users:read: Read user
            users:create: Create user
            users:update: Update user
            users:delete: Delete user
            groups:list: List groups
            groups:read: Read group
            groups:create: Create group
            groups:update: Update group
            groups:delete: Delete group
            courses:list: List courses
            courses:read: Read course
            courses:create: Create course
            courses:update: Update course
            courses:delete: Delete course
            grades:list: List grades
            grades:read: Read grade
            grades:create: Create grade
            grades:update: Update grade
            grades:delete: Delete grade
            analytics:read: Read analytics
            auth:me: Read current user
            roles:update: Update user roles
          tokenUrl: /api/v1/auth/login